backend/requirements.txt
fastapi
mysql-connector-python
uvicorn
sqlalchemy
python-dotenv